Move FilterPopover to helper plugin
