+++ packages/core/admin/admin/src/content-manager/components/AttributeFilter/Filters.js
