--- packages/core/admin/admin/src/content-manager/components/AttributeFilter/utils/formatAttribute.js
@@ -1,13 +1,0 @@
-import get from 'lodash/get';
-
-const formatAttribute = (attributeName, metaData) => {
-  const mainField = get(metaData, [attributeName, 'list', 'mainField', 'name']);
-
-  if (mainField) {
-    return `${attributeName}.${mainField}`;
-  }
-
-  return attributeName;
-};
-
-export default formatAttribute;
