--- packages/core/admin/admin/src/content-manager/components/AttributeFilter/utils/getAttributeType.js
@@ -1,13 +1,0 @@
-import get from 'lodash/get';
-
-const getAttributeType = (attributeName, contentType, metaData) => {
-  let attributeType = get(contentType, ['attributes', attributeName, 'type'], '');
-
-  if (attributeType === 'relation') {
-    attributeType = get(metaData, [attributeName, 'list', 'mainField', 'schema', 'type'], 'string');
-  }
-
-  return attributeType === 'string' ? 'text' : attributeType;
-};
-
-export default getAttributeType;
